requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-OAuthlib==0.9.6
google-cloud-bigquery==3.11.4
gunicorn==20.1.0
